create: router miner
